+++ ledger/api/character/ledger.py
@@ -1,16 +1,16 @@
 from typing import List
 
 from ninja import NinjaAPI
-from ninja.pagination import paginate
 
 from ledger.api import schema
-from ledger.api.helpers import Paginator, get_alts_queryset, get_main_character
+from ledger.api.helpers import get_alts_queryset, get_main_character
 from ledger.api.managers.ledger_manager import JournalProcess
 from ledger.hooks import get_extension_logger
 
 logger = get_extension_logger(__name__)
 
 
+# pylint: disable=duplicate-code
 class LedgerApiEndpoints:
     tags = ["CharacerLedger"]
 
@@ -20,7 +20,6 @@ class LedgerApiEndpoints:
             response={200: List[schema.CharacterLedger], 403: str},
             tags=self.tags,
         )
-        @paginate(Paginator)
         def get_character_ledger(request, character_id: int, year: int, month: int):
             if character_id == 0:
                 character_id = request.user.profile.main_character.character_id